+++ pychron/processing/tasks/tables/editors/summary_table_editor.py
@@ -62,7 +62,7 @@ class SummaryTableEditor(BaseTableEditor, ColumnSorterMixin):
         return cs
 
     def _load_widths(self):
-        p = os.path.join(paths.hidden_dir, 'summary_col_widths')
+        p = os.path.join(paths.hidden_dir, 'summary_col_widths.p')
         if os.path.isfile(p):
             with open(p, 'r') as fp:
                 try:
@@ -71,7 +71,7 @@ class SummaryTableEditor(BaseTableEditor, ColumnSorterMixin):
                     print 'load_widths', e
 
     def _dump_widths(self):
-        p = os.path.join(paths.hidden_dir, 'summary_col_widths')
+        p = os.path.join(paths.hidden_dir, 'summary_col_widths.p')
         with open(p, 'w') as fp:
             pickle.dump(self.col_widths, fp)